Split up views.py and views_firmware.py into managable chunks
No other code changes.
