+++ app/templates/metadata.html
@@ -56,7 +56,7 @@
 <p>
   This rebuilds metadata for all targets and QA groups.
 </p>
-<form method=\"get\" action="/lvfs/metadata_rebuild">
+<form method=\"get\" action="/lvfs/metadata/rebuild">
 <button class="btn btn-danger">Rebuild Metadata</button>
 </form>
 {% endif %}
